requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-multipart
pydantic>=2.0
requests
python-dotenv
streamlit
plotly
